Move mixin back in main class. Move internal variable up in definition
